size_t precision

+++ cpp/src/interval/boost_interval.hpp
@@ -294,7 +294,7 @@ public:
         return BoostInterval(boost::numeric::interval_lib::pi<BoostIntervalType>());
     }
 
-    static void set_print_precision(const int print_precision) {
+    static void set_print_precision(const size_t print_precision) {
         Number::set_print_precision(print_precision);
     }
