+++ cpp/src/interval/mpfi_interval.hpp
@@ -429,7 +429,7 @@ public:
         return pi;
     }
 
-    static void set_print_precision(const int print_precision) {
+    static void set_print_precision(const size_t print_precision) {
         Number::set_print_precision(print_precision);
     }
